requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-recaptcha==2.0.6
django-taggit==1.3.0
Pillow==8.0.1
pkg-resources==0.0.0
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pytz==2020.5
sqlparse==0.4.1


